server/pyproject.toml
[tool.poetry]
name = "Notes"
version = "0.0.1"
description = "Notes test task"
authors = ["Alexander Shcherbanyuk <@s0ul701>"]

[tool.poetry.dependencies]
python = "3.9.5"
Django = "^3.2.4"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
django-admin-sortable2 = "^1.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.1"
django-extensions = "^3.1.3"
Werkzeug = "^2.0.1"
flake8 = "^3.9.2"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
